Time complexity: O(n*m), Space complexity: O(n*m)

+++ 3619. Count Islands With Total Value Divisible by K.cpp
@@ -1,13 +1,11 @@
-#include <vector>
-
 class Solution {
 public:
-    int dfs(std::vector<std::vector<int>>& grid, int i, int j, int n, int m, std::vector<std::vector<bool>>& visited) {
+    long long dfs(std::vector<std::vector<int>>& grid, int i, int j, int n, int m, std::vector<std::vector<bool>>& visited) {
         if (i >= n || j >= m || i < 0 || j < 0) return 0;
         if (grid[i][j] == 0 || visited[i][j]) return 0;
 
         visited[i][j] = true;
-        int val = dfs(grid, i + 1, j, n, m, visited) +
+        long long val = dfs(grid, i + 1, j, n, m, visited) +
                   dfs(grid, i, j + 1, n, m, visited) +
                   dfs(grid, i - 1, j, n, m, visited) +
                   dfs(grid, i, j - 1, n, m, visited) +
@@ -25,7 +23,7 @@ public:
         for (int i = 0; i < n; ++i) {
             for (int j = 0; j < m; ++j) {
                 if (grid[i][j] != 0 && !visited[i][j]) {
-                    int val = dfs(grid, i, j, n, m, visited);
+                    long long val = dfs(grid, i, j, n, m, visited);
                     if (val % k == 0) ans++;
                 }
             }
